Replace old colorpicker with new component
